refactor: 拆分 Respond 编排职责
